pyproject.toml
[project]
name = "hci"
version = "0.1.0"
description = "The HLA Curation Interface is a web application that biocurators can use to classify HLA alleles and haplotypes."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2",
    "django-htmx>=1.23.0",
    "pydantic>=2.11.3",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "coverage>=7.8.0",
    "django-stubs>=5.2.0",
    "pyright[nodejs]>=1.1.400",
    "pytest-django>=4.11.1",
    "ruff>=0.11.5",
    "sphinx>=8.2.3",
    "types-requests>=2.32.0.20250328",
]

[tool.ruff]
line-length = 88
target-version = "py313"
exclude = [
    "migrations",
    ".mypy_cache",
    "__pycache__",
    "docs",
]

[tool.ruff.lint]
preview = true  # Use Ruff's "preview" (beta) features.
select = [
    # Basics:
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes rules
    "N",    # pep8-naming rules (enforce PEP8 naming conventions)
    "C90",  # mccabe (don't allow high cyclomatic complexity)

    # Imports:
    "I",    # isort rules (sort imports)

    # Docstrings:
    "D",    # pydocstyle (enforce style for docstrings)
    "DOC",  # pydoclint (lint docstrings)

    # flake8 plugins:
    "ANN",   # flake8-annotations (use PEP 3107 function annotations)
    "S",     # flake8-bandit (perform security-focused static analysis)
    "FBT",   # flake8-boolean-trap (don't use boolean trap antipattern)
    "B",     # flake8-bugbear (provide an opinionated set of lint rules)
    "A",     # flake8-builtins (provide a set of lint rules for builtins)
    "COM",   # flake8-commas (provide a set of lint rules for commas)
    "C4",    # flake8-comprehensions (provide a set of lint rules for comprehensions)
    "DTZ",   # flake8-datetimez (use the datetime library in a safe way)
    "T10",   # flake8-debugger (don't allow debugging code)
    "DJ",    # flake8-django (provide a set of lint rules for Django)
    "EM",    # flake8-errmsg (provide a set of rules for error messages)
    "EXE",   # flake8-executable (provide a set of lint rules for Python scripts)
    "FIX",   # flake8-fixme (don't allow "I'll do it later" comments)
    "ISC",   # flake8-implicit-string-concat (provide a set of lint rules for string concatenation)
    "LOG",   # flake8-logging (provide a set of lint rules for logging)
    "G",     # flake8-logging-format (provide a set of lint rules for the format of log strings),
    "INP",   # flake8-no-pep420 (don't allow implicit namespaces)
    "PIE",   # flake8-pie (provide a set of miscellaneous lint rules)
    "T20",   # flake8-print (don't allow print statements; use logs instead)
    "PT",    # flake8-pytest-style (provide a set of lint rules for Pytest tests)
    "Q",     # flake8-quotes (provide a set of lint rules about quotes)
    "RSE",   # flake8-raise (provide a set of lint rules for the raise statement)
    "RET",   # flake8-return (provide a set of lint rules for the return statement)
    "SLF",   # flake8-self (don't allow access to private members)
    "SIM",   # flake8-simplify (provide a set of lint rules for simplifying various expressions)
    "TID",   # flake8-tidy-imports (provide a set of lint rules for imports)
    "TD",    # flake8-todos (provide a set of lint rules for todo comments)
    "ARG",   # flake8-unused-arguments (provide a set of lint rules for unused arguments)
    "PTH",   # flake8-use-pathlib (use the new pathlib library)

    # Other:
    "PERF",  # Perflint (provide a set of lint rules for performance)
    "UP",    # pyupgrade (use new Python syntax)
    "RUF",   # Ruff-specific rules (provide a set of miscellaneous lint rules)
    "TRY",   # tryceratops (provide a set of lint rules for try-except statements)
]
ignore = [
    "COM812", # This rule clashes with Ruff.
    "G004",   # Not using f-strings in logs seems like a premature optimization.
    "RUF012", # Django models be like that.
    "RUF100", # With Ruff 0.11.5, I'm seeing false positives.
    "S314",   # While I agree with this rule, it suggests using an outdated third-party package.
    "S405",   # While I agree with this rule, it suggests using an outdated third-party package.
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = [
    "D104", # Docstrings in these files are overkill.
    "F401", # In some cases, I want to be able to import the package name rather than the individual modules.
]
"*_test.py" = [
    "S101", # We need `assert` statements in tests.
]
"conftest.py" = [  # This file configures pytest and adds a custom pytest rule.
    "ARG001",      # The unused argument is required in this case.
    "T201",        # I want to report probelms to the user.
]

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
strict_equality = true
warn_return_any = true
warn_unused_ignores = true

[tool.django-stubs]
django_settings_module = "config.settings.dev"
strict_settings = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.dev"
python_files = ["*_test.py"]
markers = [
    "unit: Make sure a single module with no other dependencies works in isolation.",
    "component: Make sure a single module, allowing necessary dependencies like the database, works.",
    "integration: Make sure multiple modules work together.",
    "contract: Make sure an external data source works.",
    "e2e: Make sure a workflow from the frontend to the database works.",
]
